requirements.txt
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.11.0
